Revert "fix : a11y - header and few things (#178)"
This reverts commit c0669d9f92dd313b7cc1413c207d24fdaddc2042.
