+++ .github/workflows/ci.yml
@@ -2,12 +2,12 @@ name: CI
 
 on:
   push:
-    branches: ['main']
+    branches: ["main"]
   pull_request:
   workflow_dispatch:
 env:
-  PUBLIC_API_URL: 'https://mj-cms.jcloud-ver-jpe.ik-server.com'
-  SENDINBLUE_API_KEY: '' # Keep this variable empty.
+  PUBLIC_API_URL: "https://mj-cms.jcloud-ver-jpe.ik-server.com"
+  SENDINBLUE_API_KEY: "" # Keep this variable empty.
 
 jobs:
   test:
@@ -18,19 +18,19 @@ jobs:
       - uses: actions/setup-node@v3
         with:
           node-version: 16.17
-          cache: 'npm'
+          cache: "npm"
           cache-dependency-path: package-lock.json
 
-      - name: 'Install dependencies'
+      - name: "Install dependencies"
         run: npm ci
 
-      - name: 'Build'
+      - name: "Build"
         run: npm run build
 
-      - name: 'Run checks'
+      - name: "Run checks"
         run: npm run check
 
-      - name: 'Run tests'
+      - name: "Run tests"
         run: npm run test:unit
 
       - name: Upload test results
