+++ src/app.css
@@ -90,7 +90,7 @@
 	/* Link */
 
 	.mj-link--base {
-		@apply text-indigo underline;
+		@apply text-indigo;
 	}
 
 	[href],
@@ -106,7 +106,7 @@
 	}
 
 	.mj-link--alt {
-		@apply text-sand no-underline;
+		@apply text-sand;
 	}
 
 	.mj-link--btn {
