+++ src/lib/components/mollecules/DesktopMenuItem/DesktopMenuItem.svelte
@@ -8,7 +8,7 @@
 		class:text-white={!isActive}
 		class:underline={isActive}
 		class:text-sand={isActive}
-		class="text-white no-underline hover:underline decoration-2 underline-offset-[8px] decoration-sand text-lg"
+		class="text-white hover:underline decoration-2 underline-offset-[8px] decoration-sand text-lg"
 		{href}
 		aria-current={isActive ? 'page' : null}><slot /></a
 	>
