+++ src/lib/components/mollecules/Header/Header.svelte
@@ -1,8 +1,70 @@
 <script lang="ts">
+	import Icon from '$lib/components/atoms/Icon/Icon.svelte';
 	import { page } from '$app/stores';
-	import DesktopMainNavigation from './DesktopMainNavigation.svelte';
-	import MobiileMainNavigation from './MobileMainNavigation.svelte';
+	import MenuMobile from '../MenuMobile/MenuMobile.svelte';
+	import DesktopMenu from '../DesktopMenu/DesktopMenu.svelte';
+	import SkipLink from '$lib/components/atoms/SkipLink/SkipLink.svelte';
+
+	export let isMenuOpen = false;
 </script>
 
-<DesktopMainNavigation activeRoute={$page.route.id} />
-<MobiileMainNavigation />
+<div class="bg-blue-dark sticky top-0 font-ptsans text-white md:px-16 p-4 z-50">
+	<SkipLink />
+	<!-- Menu Desktop -->
+	<!-- svelte-ignore a11y-no-redundant-roles -- this is the main navigation menu -->
+	<nav
+		class="lg:flex w-full top-0 items-center justify-between hidden"
+		role="navigation"
+		aria-label="menu principal"
+	>
+		<a
+			class="text-white lg:hidden xl:block"
+			title="MJ Communication &
+		transition écologique - Accueil"
+			href="/"
+		>
+			<Icon
+				role="img"
+				ariaLabel="MJ Communication &
+			transition écologique"
+				height="auto"
+				width="130px"
+				name="mjConseilFull"
+			/>
+		</a>
+
+		<a class="text-white xl:hidden md:hidden lg:block" aria-label="Vers la page d'accueil" href="/">
+			<Icon
+				role="img"
+				ariaLabel="MJ Communication &
+		transition écologique"
+				height="auto"
+				width="75px"
+				name="mjConseilFull"
+			/>
+		</a>
+
+		<DesktopMenu activeRoute={$page.route.id} />
+	</nav>
+
+	<!-- Menu Mobile -->
+	<div class="flex lg:hidden justify-between w-full">
+		<a class="text-white" aria-label="Vers la page d'accueil" href="/">
+			<Icon height="35px" width="auto" name="mjConseilFull" />
+		</a>
+		<button on:click on:click={() => (isMenuOpen = true)} aria-label="ouvrir le menu de navigation">
+			<Icon height="2rem" width="2rem" name="burger" />
+		</button>
+		<MenuMobile
+			on:clickLink={() => (isMenuOpen = false)}
+			on:clickLink
+			on:click={() => (isMenuOpen = false)}
+			on:closeModal={() => (isMenuOpen = false)}
+			on:closeModal
+			{isMenuOpen}
+		/>
+	</div>
+</div>
+
+<style>
+</style>
