--- src/lib/components/mollecules/Header/HeaderLogo.svelte
@@ -1,32 +1,0 @@
-<script>
-	import Icon from '$lib/components/atoms/Icon/Icon.svelte';
-</script>
-
-<a
-	class="text-white lg:hidden xl:block"
-	title="MJ Communication &
-transition écologique - Accueil"
-	href="/"
->
-	<Icon
-		role="img"
-		ariaLabel="MJ Communication &
-transition écologique"
-		height="auto"
-		width="130px"
-		name="mjConseilFull"
-		id="mainLogo"
-	/>
-</a>
-
-<a class="text-white xl:hidden md:hidden lg:block" aria-label="Vers la page d'accueil" href="/">
-	<Icon
-		role="img"
-		ariaLabel="MJ Communication &
-transition écologique"
-		height="auto"
-		width="75px"
-		name="mjConseilFull"
-		id="mainLogo"
-	/>
-</a>
