+++ src/lib/constants.ts
@@ -39,6 +39,3 @@ export enum SIZE {
 	REGULAR = 'regular',
 	MEDIUM = 'medium'
 }
-
-export const focusableElements =
-	'button, a, [href], input, select, textarea, [tabindex]:not([tabindex="-1"]), summary';
