+++ src/routes/+layout.svelte
@@ -1,21 +1,23 @@
-<script lang="ts">
+<script>
 	import Header from '$lib/components/mollecules/Header/Header.svelte';
 	import Footer from '$lib/components/mollecules/Footer/Footer.svelte';
 	import '../app.css';
+	import SkipLink from '$lib/components/atoms/SkipLink/SkipLink.svelte';
 
-	import { afterNavigate } from '$app/navigation';
-
-	let node: HTMLElement;
-
-	// note : this is a bit hacking ... After navigating to a page with a "subroute" (aka notre-offre/xxxx) the focus was not correctly reset
-	// this code ensure that ...
-	afterNavigate(() => {
-		node.focus();
-	});
+	let modalShown = false;
 </script>
 
-<div bind:this={node} tabindex="-1" class="relative h-full">
-	<Header />
+<Header
+	on:click={() => {
+		modalShown = true;
+	}}
+	on:clickLink={() => {
+		modalShown = false;
+	}}
+	on:closeModal={() => (modalShown = false)}
+/>
+
+<div class="relative" inert={modalShown ? true : null} class:overflow-hidden={modalShown}>
 	<slot />
 
 	<Footer />
