+++ src/routes/blog/[slug]/+page.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+	import SkipLink from '$lib/components/atoms/SkipLink/SkipLink.svelte';
 	import Tag from '$lib/components/atoms/Tag/Tag.svelte';
 	import PostCard from '$lib/components/mollecules/PostCard/PostCard.svelte';
 	import Section from '$lib/components/mollecules/Section/Section.svelte';
