+++ src/routes/espace-presse/[slug]/+page.ts
@@ -1,5 +1,5 @@
 import { getMediaResources } from '$lib/repositories/mediaResources';
-import { error } from '@sveltejs/kit';
+import { error, } from '@sveltejs/kit';
 
 import type { PageLoad } from './$types';
